chore(js): moved defineFormat and defineInterrupt to beta, api ref fixes
