+++ js/genkit/tests/generate_test.ts
@@ -19,17 +19,17 @@ import { z } from '@genkit-ai/core';
 import * as assert from 'assert';
 import { beforeEach, describe, it } from 'node:test';
 import { modelRef } from '../../ai/src/model';
-import { Genkit, genkit } from '../src/genkit';
 import {
   ProgrammableModel,
   defineEchoModel,
   defineProgrammableModel,
   runAsync,
 } from './helpers';
+import { GenkitBeta, genkit } from '../src/beta';
 
 describe('generate', () => {
   describe('default model', () => {
-    let ai: Genkit;
+    let ai: GenkitBeta;
 
     beforeEach(() => {
       ai = genkit({
@@ -145,7 +145,7 @@ describe('generate', () => {
   });
 
   describe('streaming', () => {
-    let ai: Genkit;
+    let ai: GenkitBeta;
 
     beforeEach(() => {
       ai = genkit({});
@@ -253,7 +253,7 @@ describe('generate', () => {
   });
 
   describe('config', () => {
-    let ai: Genkit;
+    let ai: GenkitBeta;
 
     beforeEach(() => {
       ai = genkit({});
@@ -303,7 +303,7 @@ describe('generate', () => {
   });
 
   describe('tools', () => {
-    let ai: Genkit;
+    let ai: GenkitBeta;
     let pm: ProgrammableModel;
 
     beforeEach(() => {
