+++ js/genkit/typedoc.json
@@ -1,20 +1,21 @@
 {
   "entryPoints": [
     "src/index.ts",
+    "src/beta.ts",
+    "src/client/index.ts",
+    "src/model.ts",
+    "src/tool.ts",
+    "src/retriever.ts",
+    "src/reranker.ts",
+    "src/plugin.ts",
+    "src/middleware.ts",
     "src/registry.ts",
     "src/tracing.ts",
     "src/logging.ts",
     "src/schema.ts",
-    "src/retriever.ts",
-    "src/reranker.ts",
     "src/embedder.ts",
     "src/evaluator.ts",
-    "src/model.ts",
-    "src/middleware.ts",
     "src/extract.ts",
-    "src/testing.ts",
-    "src/tool.ts",
-    "src/plugin.ts",
-    "src/client/index.ts"
+    "src/testing.ts"
   ]
 }
